refactor: Move ui specific methods to ui
