+++ src/lib/ui/SupportedDeviceTable/SupportedDevice.stories.tsx
@@ -1,7 +1,7 @@
 import { Button, Dialog, DialogActions } from '@mui/material'
 import type { Meta, StoryObj } from '@storybook/react'
 
-import { useToggle } from 'lib/use-toggle.js'
+import { useToggle } from 'lib/ui/use-toggle.js'
 
 import { SupportedDeviceTable } from './SupportedDeviceTable.js'
