+++ src/lib/ui/use-interval.ts
